@@ -34,9 +34,8 @@
     "event-emitter": "^0.3.4",
     "foundation-apps": "git+https://github.com/zurb/foundation-apps.git",
     "git-rev-sync": "^1.4.0",
-    "graphenejs-lib": "^0.4.6",
-    "graphenejs-ws": "^0.1.3",
-    "highcharts": "^4.2.3",
+    "graphenejs-lib": "^0.4.10",
+    "highcharts": "4.2.6",
     "history": "1.13.1",
     "html-loader": "^0.3.0",
     "immutable": "^3.8.1",
@@ -50,11 +49,11 @@
     "object-assign": "^4.0.1",
     "peerjs": "^0.3.14",
     "perfect-scrollbar": "^0.6.11",
-    "react": "^15.2.0",
-    "react-addons-css-transition-group": "^15.1.0",
-    "react-dom": "^15.2.0",
+    "react": "^15.3.2",
+    "react-addons-css-transition-group": "^15.3.2",
+    "react-dom": "^15.3.2",
     "react-foundation-apps": "git+https://github.com/valzav/react-foundation-apps.git",
-    "react-highcharts": "^8.4.1",
+    "react-highcharts": "^10.0.0",
     "react-interpolate-component": "^0.9.0",
     "react-intl": "^2.0.1",
     "react-json-inspector": "^7.0.0",
@@ -87,6 +86,7 @@
     "jest-cli": "~0.4.1",
     "json-loader": "^0.5.3",
     "mocha": "^2.5.3",
+    "node-libs-browser": "^1.0.0",
     "node-sass": "^3.4.1",
     "postcss-loader": "^0.8.2",
     "react-transform-hmr": "^1.0.4",
