@@ -12,29 +12,29 @@ try {
 }
 
 /**
-    This should only be applied to a BTS 1.0 export file taken on the 
+    This should only be applied to a BTS 1.0 export file taken on the
     discontinued chain. Any public key string or address (all 5 formats) carried
     over to the BTS 2.0 genesis block will be in this filter.
 
     Their may be some false positives but no false negatives.
 */
 export default class GenesisFilter {
-    
+
     /** or call this.init */
     constructor(bloom_buffer) {
         if( ! bloom_buffer ) return
         this.bloom_buffer = bloom_buffer
         this.bits_in_filter = bloom_buffer.length * 8 // 8388608 (test data)
     }
-    
+
     /** Was a bloom file deployed?  This does not try to load it from the server. */
     isAvailable() { return bts_genesiskeys_bloom_url !== undefined }
-    
+
     init(done) {
         if( this.bloom_buffer ) { done(); return }
         if( ! this.isAvailable() )
             throw new Error("Genesis bloom file was not deployed")
-        
+
         var xhr = new XMLHttpRequest
         // firefox 40 did not allow the blob url but ff 41.0.2 did
         xhr.responseType = "blob"
@@ -54,7 +54,7 @@ export default class GenesisFilter {
         xhr.open("GET", bts_genesiskeys_bloom_url)
         xhr.send()
     }
-    
+
     inGenesis(pubkey_or_address) {
         if( ! this.bloom_buffer ) throw new Error("Call init() first")
         for(var hashes = 0; hashes < 3; hashes++) {
@@ -72,7 +72,7 @@ export default class GenesisFilter {
         }
         return true
     }
-    
+
     filter( account_keys, status ) {
         if( ! this.isAvailable() ) {
             console.log("WARN: Missing bloom filter for BTS 0.9.x wallets")
@@ -102,10 +102,10 @@ export default class GenesisFilter {
                             status({ error: "missing_public_keys" })
                             return
                         }
-                        var key = keys.public_keys[k]
-                        if( /^GPH/.test(key) ) key = "BTS" + key.substring(3)
-                        if(this.inGenesis( key )) continue
-                        var addresses = key.addresses(key, 'BTS')
+                        var currentKey = keys.public_keys[k]
+                        if( /^GPH/.test(currentKey) ) currentKey = "BTS" + currentKey.substring(3)
+                        if(this.inGenesis( currentKey )) continue
+                        var addresses = key.addresses(currentKey, 'BTS')
                         var addy_found = false
                         for(var i = 0; i < addresses.length; i++) {
                             if(this.inGenesis( addresses[i] )) {
@@ -138,5 +138,5 @@ export default class GenesisFilter {
             }
         })
     }
-    
+
 }
