@@ -16,7 +16,7 @@ class SettingsStore {
 
         this.defaultSettings = Immutable.Map({
             locale: "en",
-            connection: "wss://bitshares.openledger.info/ws",
+            apiServer: "wss://bitshares.openledger.info/ws",
             faucet_address: "https://bitshares.openledger.info",
             unit: CORE_ASSET,
             showSettles: false,
@@ -31,7 +31,7 @@ class SettingsStore {
             "MKR", "OPEN.MKR", "BTS", "OPEN.ETH", "ICOO", "BTC", "OPEN.LISK",
             "OPEN.STEEM", "OPEN.DAO", "PEERPLAYS", "USD", "CNY", "BTSR", "OBITS",
             "OPEN.DGD", "EUR", "TRADE.BTC", "CASH.BTC", "GOLD", "SILVER",
-            "OPEN.USDT", "OPEN.EURT", "OPEN.BTC", "CADASTRAL"
+            "OPEN.USDT", "OPEN.EURT", "OPEN.BTC", "CADASTRAL", "BLOCKPAY"
         ];
 
         this.preferredBases = Immutable.List([CORE_ASSET, "OPEN.BTC", "USD", "CNY", "BTC"]);
@@ -53,6 +53,16 @@ class SettingsStore {
 
         // If you want a default value to be translated, add the translation to settings in locale-xx.js
         // and use an object {translate: key} in the defaults array
+        let apiServer = [
+            {url: "wss://bitshares.openledger.info/ws", location: "Nuremberg, Germany"},
+            {url: "wss://bit.btsabc.org/ws", location: "Hong Kong"},
+            {url: "wss://bts.transwiser.com/ws", location: "Hangzhou, China"},
+            {url: "wss://bitshares.dacplay.org:8089/ws", location:  "Hangzhou, China"},
+            {url: "wss://openledger.hk/ws", location: "Hong Kong"},
+            {url: "wss://secure.freedomledger.com/ws", location: "Toronto, Canada"},
+            {url: "wss://testnet.bitshares.eu/ws", location: "Public Testnet Server (Frankfurt, Germany)"}
+        ];
+
         let defaults = {
             locale: [
                 "en",
@@ -63,12 +73,7 @@ class SettingsStore {
                 "es",
                 "tr"
             ],
-            connection: [
-                "wss://bitshares.openledger.info/ws",
-                "wss://bitshares.dacplay.org:8089/ws",
-                "wss://dele-puppy.com/ws",
-                "wss://valen-tin.fr:8090/ws"
-            ],
+            apiServer: [],
             unit: [
                 CORE_ASSET,
                 "USD",
@@ -122,7 +127,52 @@ class SettingsStore {
 
         this.starredAccounts = Immutable.Map(ss.get("starredAccounts"));
 
-        this.defaults = merge({}, defaults, ss.get("defaults_v1"));
+        let savedDefaults = ss.get("defaults_v1", {});
+        this.defaults = merge({}, defaults, savedDefaults);
+
+        (savedDefaults.connection || []).forEach(api => {
+            let hasApi = false;
+            if (typeof api === "string") {
+                api = {url: api, location: null};
+            }
+            apiServer.forEach(server => {
+                if (server.url === api.url) {
+                    hasApi = true;
+                }
+            });
+
+            if (!hasApi) {
+                this.defaults.apiServer.push(api);
+            }
+        });
+
+        (savedDefaults.apiServer || []).forEach(api => {
+            let hasApi = false;
+            if (typeof api === "string") {
+                api = {url: api, location: null};
+            }
+            this.defaults.apiServer.forEach(server => {
+                if (server.url === api.url) {
+                    hasApi = true;
+                }
+            });
+
+            if (!hasApi) {
+                this.defaults.apiServer.push(api);
+            }
+        });
+
+        for (let i = apiServer.length - 1; i >= 0; i--) {
+            let hasApi = false;
+            this.defaults.apiServer.forEach(api => {
+                if (api.url === apiServer[i].url) {
+                    hasApi = true;
+                }
+            });
+            if (!hasApi) {
+                this.defaults.apiServer.unshift(apiServer[i]);
+            }
+        }
 
         this.viewSettings = Immutable.Map(ss.get("viewSettings_v1"));
 
@@ -213,13 +263,16 @@ class SettingsStore {
     }
 
     onAddWS(ws) {
-        this.defaults.connection.push(ws);
+        if (typeof ws === "string") {
+            ws = {url: ws, location: null};
+        }
+        this.defaults.apiServer.push(ws);
         ss.set("defaults_v1", this.defaults);
     }
 
     onRemoveWS(index) {
-        if (index !== 0) { // Prevent removing the default connection
-            this.defaults.connection.splice(index, 1);
+        if (index !== 0) { // Prevent removing the default apiServer
+            this.defaults.apiServer.splice(index, 1);
             ss.set("defaults_v1", this.defaults);
         }
     }
