@@ -62,13 +62,13 @@ import Brainkey from "./components/Wallet/Brainkey";
 import AccountRefsStore from "stores/AccountRefsStore";
 import Help from "./components/Help";
 import InitError from "./components/InitError";
+import SyncError from "./components/SyncError";
 import BrowserSupportModal from "./components/Modal/BrowserSupportModal";
 import createBrowserHistory from "history/lib/createHashHistory";
 import {IntlProvider} from "react-intl";
 import intlData from "./components/Utility/intlData";
 import connectToStores from "alt/utils/connectToStores";
 import Chat from "./components/Chat/ChatWrapper";
-import Icon from "./components/Icon/Icon";
 import Translate from "react-translate-component";
 
 require("./components/Utility/Prototypes"); // Adds a .equals method to Array for use in shouldComponentUpdate
@@ -89,6 +89,7 @@ class App extends React.Component {
             theme: SettingsStore.getState().settings.get("themes"),
             disableChat: SettingsStore.getState().settings.get("disableChat", true),
             showChat: SettingsStore.getState().viewSettings.get("showChat", false),
+            dockedChat: SettingsStore.getState().viewSettings.get("dockedChat", false),
             isMobile: false
         };
     }
@@ -177,6 +178,12 @@ class App extends React.Component {
             });
         }
 
+        if (viewSettings.get("dockedChat") !== this.state.dockedChat) {
+            this.setState({
+                dockedChat: viewSettings.get("dockedChat")
+            });
+        }
+
     }
 
     // /** Non-static, used by passing notificationSystem via react Component refs */
@@ -186,7 +193,7 @@ class App extends React.Component {
     // }
 
     render() {
-        let {disableChat, isMobile, showChat} = this.state;
+        let {disableChat, isMobile, showChat, dockedChat} = this.state;
 
         let content = null;
 
@@ -194,17 +201,7 @@ class App extends React.Component {
 
         if (this.state.syncFail) {
             content = (
-                <div className="grid-frame vertical">
-                    <div className="grid-container text-center" style={{paddingTop: "5rem"}}>
-
-                        <h2><Translate content="sync_fail.title" /></h2>
-                        <br />
-                        <p><Translate content="sync_fail.sub_text_1" /></p>
-                        <p><Translate content="sync_fail.sub_text_2" /></p>
-                        <Icon name="clock" size="5x"/>
-                    </div>
-
-                </div>
+                <SyncError />
             );
         } else if (this.state.loading) {
             content = <div className="grid-frame vertical"><LoadingIndicator /></div>;
@@ -220,7 +217,13 @@ class App extends React.Component {
                             {this.props.children}
                         </div>
                         <div className="grid-block shrink" style={{overflow: "hidden"}}>
-                            {isMobile ? null : <Chat showChat={showChat} disable={disableChat} footerVisible={showFooter}/>}
+                            {isMobile ? null :
+                                <Chat
+                                    showChat={showChat}
+                                    disable={disableChat}
+                                    footerVisible={showFooter}
+                                    dockedChat={dockedChat}
+                                />}
 
                         </div>
                     </div>
@@ -280,11 +283,11 @@ class Auth extends React.Component {
 
 let willTransitionTo = (nextState, replaceState, callback) => {
 
-    let connectionString = SettingsStore.getSetting("connection");
+    let connectionString = SettingsStore.getSetting("apiServer");
 
     if (nextState.location.pathname === "/init-error") {
 
-        return Apis.reset(connectionString).init_promise
+        return Apis.reset(connectionString, true).init_promise
         .then(() => {
             var db = iDB.init_instance(window.openDatabase ? (shimIndexedDB || indexedDB) : indexedDB).init_promise;
             return db.then(() => {
@@ -299,7 +302,7 @@ let willTransitionTo = (nextState, replaceState, callback) => {
         });
 
     }
-    Apis.instance(connectionString).init_promise.then(() => {
+    Apis.instance(connectionString, true).init_promise.then(() => {
         var db;
         try {
             db = iDB.init_instance(window.openDatabase ? (shimIndexedDB || indexedDB) : indexedDB).init_promise;
