@@ -46,6 +46,18 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
             receive_address: null,
             url: props.url || urls[props.gateway]
         };
+
+        this._copy = this._copy.bind(this);
+        document.addEventListener("copy", this._copy);
+    }
+
+    _copy(e) {
+        try {
+            e.clipboardData.setData("text/plain", this.state.clipboardText);
+            e.preventDefault();
+        } catch(err) {
+            console.error(err);
+        }
     }
 
     componentWillMount() {
@@ -65,7 +77,7 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
         };
 
         let body_string = JSON.stringify(body);
- 
+
         fetch( this.state.url + '/simple-api/initiate-trade', {
             method:'post',
             headers: new Headers( { "Accept": "application/json", "Content-Type":"application/json" } ),
@@ -101,9 +113,19 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
     onWithdraw() {
         ZfApi.publish(this.getWithdrawModalId(), "open");
     }
-    
+
+    toClipboard(clipboardText) {
+        try {
+            this.setState({clipboardText}, () => {
+                document.execCommand("copy");
+            });
+        } catch(err) {
+            console.error(err);
+        }
+    }
+
     render() {
-		
+
         let emptyRow = <div style={{display:"none", minHeight: 150}}></div>;
         if( !this.props.account || !this.props.issuer_account || !this.props.receive_asset )
             return emptyRow;
@@ -128,7 +150,7 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
             if (!has_nonzero_balance)
                 return emptyRow;
         }
-        
+
         // let account_balances = account_balances_object.toJS();
         // let asset_types = Object.keys(account_balances);
         // if (asset_types.length > 0) {
@@ -144,7 +166,7 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
             let account_name = this.props.account.get('name');
             receive_address = this.deposit_address_cache.getCachedInputAddress(this.props.gateway, account_name, this.props.deposit_coin_type, this.props.receive_coin_type);
         }
-        
+
         if( !receive_address ) {
             this.requestDepositAddress();
             return emptyRow;
@@ -159,10 +181,13 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
         // }
         // else
         // {
+        let clipboardText = "";
+        let memoText;
         if (this.props.deposit_account)
         {
-            deposit_address_fragment = (<span><code>{this.props.deposit_account}</code></span>);
-            deposit_memo = <code>{this.props.receive_coin_type + ':' + this.props.account.get('name')}</code>
+            deposit_address_fragment = (<span>{this.props.deposit_account}</span>);
+            clipboardText = this.props.receive_coin_type + ':' + this.props.account.get('name');
+            deposit_memo = <span>{clipboardText}</span>;
             var withdraw_memo_prefix = this.props.deposit_coin_type + ':';
         }
         else
@@ -170,20 +195,24 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
             if (receive_address.memo)
             {
                 // This is a client that uses a deposit memo (like ethereum), we need to display both the address and the memo they need to send
-                deposit_address_fragment = (<span><code>{receive_address.address}</code><br />with memo <code>{receive_address.memo}</code></span>);
+                memoText = receive_address.memo;
+                clipboardText = receive_address.address;
+                deposit_address_fragment = (<span>{receive_address.address}</span>);
+                deposit_memo = <span>{receive_address.memo}</span>;
             }
             else
             {
                 // This is a client that uses unique deposit addresses to select the output
-                deposit_address_fragment = (<span><code>{receive_address.address}</code></span>);
+                clipboardText = receive_address.address;
+                deposit_address_fragment = (<span>{receive_address.address}</span>);
             }
             var withdraw_memo_prefix = '';
         }
 
         if (this.props.action === "deposit") {
             return (
-                <div className="grid-block no-padding no-margin">
-                    <div className="grid-content shrink" style={{paddingRight: 40}}>
+                <div className="grid-block no-padding no-margin test">
+                    <div style={{paddingRight: 40, paddingBottom: 20}}>
                         <Translate component="h4" content="gateway.deposit_summary" />
                         <table style={{width: "inherit"}} className="table">
                             <tbody>
@@ -218,19 +247,20 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
                             <table className="table">
                                 <tbody>
                                     <tr>
-                                        <td></td>
-                                        <td style={{textAlign: "right"}}>{deposit_address_fragment}</td>
+                                        <td colSpan="2" style={{textAlign: "left"}}>{deposit_address_fragment}</td>
                                     </tr>
                                     {deposit_memo ? (
                                     <tr>
                                         <td>memo:</td>
-                                        <td style={{textAlign: "right"}}>{deposit_memo}</td>
+                                        <td style={{textAlign: "left"}}>{deposit_memo}</td>
                                     </tr>) : null}
                                 </tbody>
                             </table>
-                            <div style={{paddingTop: 10}}>
-                                <button className={"button"} style={{width: "100%"}} onClick={this.requestDepositAddress.bind(this)}><Translate content="gateway.generate_new" /></button>
-                            </div>    
+                            <div className="button-group" style={{paddingTop: 10}}>
+                                {deposit_address_fragment ? <div className="button" onClick={this.toClipboard.bind(this, clipboardText)}>Copy address</div> : null}
+                                {memoText ? <div className="button" onClick={this.toClipboard.bind(this, memoText)}>Copy memo</div> : null}
+                                <button className={"button"} onClick={this.requestDepositAddress.bind(this)}><Translate content="gateway.generate_new" /></button>
+                            </div>
                         </div>
                     </div>
                 </div>
@@ -238,7 +268,7 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
         } else {
             return (
                 <div className="grid-block no-padding no-margin">
-                    <div className="grid-content shrink" style={{paddingRight: 40}}>
+                    <div style={{paddingRight: 40, paddingBottom: 20}}>
                         <Translate component="h4" content="gateway.withdraw_summary" />
                         <table style={{width: "inherit"}} className="table">
                             <tbody>
@@ -275,7 +305,7 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
                         <div style={{padding: "10px 0", fontSize: "1.1rem", fontWeight: "bold"}}>
                             <div style={{paddingTop: 10}}>
                                 <button className={"button success"} style={{width: "100%"}} onClick={this.onWithdraw.bind(this)}><Translate content="gateway.withdraw_now" /> </button>
-                            </div> 
+                            </div>
                         </div>
                     </div>
                     <Modal id={withdraw_modal_id} overlay={true}>
@@ -297,7 +327,7 @@ export default class BlockTradesGatewayDepositRequest extends React.Component {
                                 memo_prefix={withdraw_memo_prefix}
                                 modal_id={withdraw_modal_id} />
                         </div>
-                    </Modal>                    
+                    </Modal>
                 </div>
             );
         }
