@@ -155,6 +155,7 @@ class Exchange extends React.Component {
                 atr: false,
                 ema: false
             }),
+            buySellTop: ws.get("buySellTop", true),
             buyFeeAssetIdx: ws.get("buyFeeAssetIdx", 0) ,
             sellFeeAssetIdx: ws.get("sellFeeAssetIdx", 0),
             indicatorSettings: ws.get("indicatorSettings") || {
@@ -1122,6 +1123,16 @@ class Exchange extends React.Component {
         };
     }
 
+    _toggleBuySellPosition() {
+        this.setState({
+            buySellTop: !this.state.buySellTop
+        });
+
+        SettingsActions.changeViewSetting({
+            buySellTop: !this.state.buySellTop
+        });
+    }
+
     render() {
         let { currentAccount, linkedAccounts, limit_orders, call_orders, totalCalls, activeMarketHistory,
             totalBids, flat_asks, flat_bids, flat_calls, invertedCalls, bids, asks, starredMarkets,
@@ -1130,7 +1141,7 @@ class Exchange extends React.Component {
 
         let {buyAmount, buyPrice, buyTotal, sellAmount, sellPrice, sellTotal, leftOrderBook,
             displayBuyPrice, displaySellPrice, buyDiff, sellDiff, indicators, indicatorSettings,
-            width, height} = this.state;
+            width, height, buySellTop} = this.state;
 
         let base = null, quote = null, accountBalance = null, quoteBalance = null, baseBalance = null, coreBalance = null,
             quoteSymbol, baseSymbol, settlementPrice = null, squeezePrice = null,
@@ -1275,10 +1286,14 @@ class Exchange extends React.Component {
         let buyForm = (
             <BuySell
                 smallScreen={smallScreen}
-                style={!smallScreen && !leftOrderBook ? {minHeight: 266} : null}
                 isOpen={this.state.buySellOpen}
                 onToggleOpen={this._toggleOpenBuySell.bind(this)}
-                className={cnames("small-12 no-padding middle-content", {disabled: isNullAccount}, leftOrderBook || smallScreen ? "medium-6" : "medium-6 xlarge-4", this.state.flipBuySell ? "order-2 sell-form" : "order-1 buy-form")}
+                className={cnames(
+                    "small-12 no-padding middle-content",
+                    {disabled: isNullAccount},
+                    leftOrderBook || smallScreen ? "medium-6" : "medium-6 xlarge-4",
+                    this.state.flipBuySell ? `order-${buySellTop ? 2 : 5} sell-form` : `order-${buySellTop ? 1 : 4} buy-form`
+                )}
                 type="bid"
                 amount={buyAmount}
                 price={displayBuyPrice}
@@ -1303,16 +1318,21 @@ class Exchange extends React.Component {
                 onChangeFeeAsset={this.onChangeFeeAsset.bind(this, "buy")}
                 isPredictionMarket={base.getIn(["bitasset", "is_prediction_market"])}
                 onFlip={!this.state.flipBuySell ? this._flipBuySell.bind(this) : null}
+                onTogglePosition={this._toggleBuySellPosition.bind(this)}
             />
         );
 
         let sellForm = (
             <BuySell
                 smallScreen={smallScreen}
-                style={!smallScreen && !leftOrderBook ? {minHeight: 266} : null}
                 isOpen={this.state.buySellOpen}
                 onToggleOpen={this._toggleOpenBuySell.bind(this)}
-                className={cnames("small-12 no-padding middle-content", {disabled: isNullAccount}, leftOrderBook || smallScreen ? "medium-6" : "medium-6 xlarge-4", this.state.flipBuySell ? "order-1 buy-form" : "order-2 sell-form")}
+                className={cnames(
+                    "small-12 no-padding middle-content",
+                    {disabled: isNullAccount},
+                    leftOrderBook || smallScreen ? "medium-6" : "medium-6 xlarge-4",
+                    this.state.flipBuySell ? `order-${buySellTop ? 1 : 4} buy-form` : `order-${buySellTop ? 2 : 4} sell-form`
+                )}
                 type="ask"
                 amount={sellAmount}
                 price={displaySellPrice}
@@ -1358,7 +1378,7 @@ class Exchange extends React.Component {
                 moveOrderBook={this._moveOrderBook.bind(this)}
                 flipOrderBook={this.props.viewSettings.get("flipOrderBook")}
                 marketReady={marketReady}
-                wrapperClass="order-3 xlarge-order-4"
+                wrapperClass={`order-${buySellTop ? 3 : 1} xlarge-order-${buySellTop ? 4 : 1}`}
             />
         );
 
@@ -1587,7 +1607,11 @@ class Exchange extends React.Component {
                                 {limit_orders.size > 0 && base && quote ? (
                                 <MyOpenOrders
                                     smallScreen={this.props.smallScreen}
-                                    className={cnames({disabled: isNullAccount}, !smallScreen && !leftOrderBook ? "medium-6 xlarge-4" : "", "small-12 medium-6 no-padding align-spaced ps-container middle-content order-4")}
+                                    className={cnames(
+                                        {disabled: isNullAccount},
+                                        !smallScreen && !leftOrderBook ? "medium-6 xlarge-4" : "",
+                                        `small-12 medium-6 no-padding align-spaced ps-container middle-content order-${buySellTop ? 4 : 6}`
+                                    )}
                                     key="open_orders"
                                     orders={limit_orders}
                                     currentAccount={currentAccount.get("id")}
