@@ -18,9 +18,9 @@ class InitError extends React.Component {
     static getPropsFromStores() {
         return {
             rpc_connection_status: BlockchainStore.getState().rpc_connection_status,
-            apis: SettingsStore.getState().defaults.connection,
-            connection: SettingsStore.getState().settings.get("connection"),
-            defaultConnection: SettingsStore.getState().defaultSettings.get("connection"),
+            apis: SettingsStore.getState().defaults.apiServer,
+            apiServer: SettingsStore.getState().settings.get("apiServer"),
+            defaultConnection: SettingsStore.getState().defaultSettings.get("apiServer"),
         }
     }
 
@@ -30,8 +30,8 @@ class InitError extends React.Component {
     }
 
     onChangeWS(e) {
-        SettingsActions.changeSetting({setting: "connection", value: e.target.value });
-        Apis.reset(e.target.value);
+        SettingsActions.changeSetting({setting: "apiServer", value: e.target.value });
+        Apis.reset(e.target.value, true);
     }
 
     onReloadClick(e) {
@@ -46,15 +46,13 @@ class InitError extends React.Component {
     }
 
     onReset() {
-        SettingsActions.changeSetting({setting: "connection", value: this.props.defaultConnection });
+        SettingsActions.changeSetting({setting: "apiServer", value: this.props.defaultConnection });
         SettingsActions.clearSettings();
     }
 
     render() {
-        console.log("-- InitError.render -->", this.props);
-
         let options = this.props.apis.map(entry => {
-            return <option key={entry} value={entry}>{entry}</option>;
+            return <option key={entry.url} value={entry.url}>{entry.location || entry.url} {entry.location ? `(${entry.url})` : null}</option>;
         });
 
         return (
@@ -65,11 +63,11 @@ class InitError extends React.Component {
                         <Translate component="h3" content={`init_error.title`} />
                         <br/>
                         <section className="block-list">
-                            <header><Translate component="span" content={`settings.connection`} /></header>
+                            <header><Translate component="span" content={`settings.apiServer`} /></header>
                             <ul>
                                 <li className="with-dropdown">
 
-                                    <select onChange={this.onChangeWS.bind(this)} value={this.props.connection}>
+                                    <select onChange={this.onChangeWS.bind(this)} value={this.props.apiServer}>
                                         {options}
                                     </select>
 
