@@ -42,15 +42,12 @@ export default class BrainkeyInput extends Component {
         }
     }
 
-    render() {
-        if (this.state.loading || !dictionary_set) {
-            return <div style={{padding: 20}}>Fetching dictionary....</div>
-        }
-        var spellcheck_words = this.state.brnkey.split(" ")
-        var checked_words = []
+    _checkBrainKey() {
+        let spellcheck_words = this.state.brnkey.split(" ")
+        let checked_words = []
         spellcheck_words.forEach( (word, i) => {
             if(word === "") return
-            var spellcheckword = word.toLowerCase()
+            let spellcheckword = word.toLowerCase()
             spellcheckword = spellcheckword.match(/[a-z]+/) //just spellcheck letters
             if(spellcheckword === null || dictionary_set.has(spellcheckword[0]))
                 checked_words.push(<span key={i} style={{padding: '1px', margin: '1px'}}>{word}</span>)
@@ -58,12 +55,14 @@ export default class BrainkeyInput extends Component {
                 checked_words.push(<MisspelledWord key={i}>{word}</MisspelledWord>)
         })
         // this.ready = checked_words.length > 0
-        var word_count_label
-        var warn = true
+        let word_count_label;
+        let warn = true;
+        let valid = true;
         if(checked_words.length > 0) {
-            if(this.state.brnkey.length < 50)
-                word_count_label = this.state.brnkey.length + " characters (50 minimum)"
-            else {
+            if(this.state.brnkey.length < 50) {
+                word_count_label = this.state.brnkey.length + " characters (50 minimum)";
+                valid = false;
+            } else {
                 if(checked_words.length < 16)
                     word_count_label = checked_words.length + " words (16 recommended)"
                 else {
@@ -72,10 +71,23 @@ export default class BrainkeyInput extends Component {
                 }
             }
         }
+
+        return {warn, valid, word_count_label, checked_words};
+    }
+
+    render() {
+        if (this.state.loading || !dictionary_set) {
+            return <div style={{padding: 20}}>Fetching dictionary....</div>
+        }
+
+        let {warn, word_count_label, checked_words} = this._checkBrainKey();
+
         return (
             <span className="">
                 <div>
-                    <textarea onChange={this.formChange.bind(this)}
+                    <textarea
+                        tabIndex={this.props.tabIndex || 1}
+                        onChange={this.formChange.bind(this)}
                         value={this.state.brnkey} id="brnkey"
                         style={{height: 100, minWidth: 450}} />
                     <div style={{textAlign: "left"}} className="grid-content no-padding no-overflow">{ checked_words }</div>
@@ -91,7 +103,8 @@ export default class BrainkeyInput extends Component {
     }
 
     formChange(event) {
-        var {id, value} = event.target
+        var {id, value} = event.target;
+        let {valid} = this._checkBrainKey();
         var state = {}
         state[id] = value
         if(id === "brnkey") {
@@ -100,7 +113,11 @@ export default class BrainkeyInput extends Component {
             state.check_digits = brnkey.length < 50 ? null :
                 hash.sha1(brnkey).toString('hex').substring(0,4)
         }
-        this.setState(state)
+        this.setState(state);
+
+        if (this.props.errorCallback) {
+            this.props.errorCallback(valid);
+        }
     }
 }
 
